+++ NMSIS/DSP/Source/BasicMathFunctions/riscv_scale_f16.c
@@ -93,13 +93,13 @@ void riscv_scale_f16(
     /* C = A * scale */
 
     /* Scale input and store result in destination buffer. */
-    *pDst++ = (*pSrc++) * scale;
+    *pDst++ = (_Float16)(*pSrc++) * (_Float16)scale;
 
-    *pDst++ = (*pSrc++) * scale;
+    *pDst++ = (_Float16)(*pSrc++) * (_Float16)scale;
 
-    *pDst++ = (*pSrc++) * scale;
+    *pDst++ = (_Float16)(*pSrc++) * (_Float16)scale;
 
-    *pDst++ = (*pSrc++) * scale;
+    *pDst++ = (_Float16)(*pSrc++) * (_Float16)scale;
 
     /* Decrement loop counter */
     blkCnt--;
@@ -120,7 +120,7 @@ void riscv_scale_f16(
     /* C = A * scale */
 
     /* Scale input and store result in destination buffer. */
-    *pDst++ = (*pSrc++) * scale;
+    *pDst++ = (_Float16)(*pSrc++) * (_Float16)scale;
 
     /* Decrement loop counter */
     blkCnt--;
