+++ NMSIS/DSP/Source/FastMathFunctions/riscv_vinverse_f16.c
@@ -47,7 +47,7 @@ void riscv_vinverse_f16(
    while (blkCnt > 0U)
    {
       
-      *pDst++ = 1.0 / *pSrc++;
+      *pDst++ = 1.0f16 / (_Float16)*pSrc++;
   
       /* Decrement loop counter */
       blkCnt--;
